fix: refactored getBlocks script and renamed to syncBlocks. Fixed tx validation for genesis block
